Accept group:version:kind argument for k8sResourcePrefix descriptor
