+++ frontend/packages/operator-lifecycle-manager/src/components/operand/const.ts
@@ -1,6 +1,4 @@
-import * as _ from 'lodash';
 import { JSONSchema6 } from 'json-schema';
-import { SpecCapability } from '../descriptors/types';
 import { SchemaType } from '@console/shared/src/components/dynamic-form';
 
 export const YAML_HELP_TEXT =
@@ -32,17 +30,6 @@ export const DEFAULT_K8S_SCHEMA: JSONSchema6 = {
   },
 };
 
-export const REGEXP_K8S_RESOURCE_CAPABILITY = _.escapeRegExp(SpecCapability.k8sResourcePrefix);
-export const REGEXP_FIELD_DEPENDENCY_CAPABILITY = _.escapeRegExp(SpecCapability.fieldDependency);
-export const REGEXP_SELECT_CAPABILITY = _.escapeRegExp(SpecCapability.select);
-
-export const REGEXP_K8S_RESOURCE_SUFFIX = new RegExp(
-  `^${REGEXP_K8S_RESOURCE_CAPABILITY}(?:core~v1~)?(.*)$`,
-);
-export const REGEXP_SELECT_OPTION = new RegExp(`${REGEXP_SELECT_CAPABILITY}(.*)$`);
-export const REGEXP_FIELD_DEPENDENCY_PATH_VALUE = new RegExp(
-  `^${REGEXP_FIELD_DEPENDENCY_CAPABILITY}([^:]*):(.*)$`,
-);
 export const HIDDEN_UI_SCHEMA = {
   'ui:widget': 'hidden',
   'ui:options': { label: false },
